requirements.txt
Django==1.8
django-grappelli==2.7.1
django-nose==1.4.1
django-redis==4.0.0
django-rest-framework-mongoengine==2.1.0
django-storages==1.1.8
djangorestframework==3.1.1
hashids==1.1.0
mongodbforms==0.3
mongoengine==0.10.0
nose==1.3.7
Pillow==3.0.0
psycopg2==2.6
pymongo==3.1
pytz==2015.4
redis==2.10.5
wheel==0.24.0
